refactor(l1): move l1 logic into separate crate
