+++ crates/l1/src/state/cache.rs
