+++ crates/tempo-zone/src/l1.rs
@@ -1,0 +1,6 @@
+//! Re-exports for L1 subscription and deposit queue types.
+
+pub use zone_l1::{
+    Deposit, DepositQueue, EnabledToken, EncryptedDeposit, L1BlockDeposits, L1Deposit,
+    L1PortalEvents, L1SequencerEvent, L1Subscriber, L1SubscriberConfig, PreparedL1Block,
+};
